Move the backup models into their own file.
