+++ api/v1beta1/foundationdbrestore_types.go
@@ -1,0 +1,63 @@
+/*
+ * foundationdbbackup_types.go
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2020 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package v1beta1
+
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+// +kubebuilder:object:root=true
+// +kubebuilder:resource:shortName=fdbrestore
+// +kubebuilder:subresource:status
+
+// FoundationDBRestore is the Schema for the FoundationDB Restore API
+type FoundationDBRestore struct {
+	metav1.TypeMeta   `json:",inline"`
+	metav1.ObjectMeta `json:"metadata,omitempty"`
+
+	Spec   FoundationDBRestoreSpec   `json:"spec,omitempty"`
+	Status FoundationDBRestoreStatus `json:"status,omitempty"`
+}
+
+// +kubebuilder:object:root=true
+
+// FoundationDBRestoreList contains a list of FoundationDBRestore objects.
+type FoundationDBRestoreList struct {
+	metav1.TypeMeta `json:",inline"`
+	metav1.ListMeta `json:"metadata,omitempty"`
+	Items           []FoundationDBRestore `json:"items"`
+}
+
+// FoundationDBRestoreSpec describes the desired state of the backup for a cluster.
+type FoundationDBRestoreSpec struct {
+	// DestinationClusterName provides the name of the cluster that the data is
+	// being restored into.
+	DestinationClusterName string `json:"destinationClusterName"`
+
+	// BackupURL provides the URL for the backup.
+	BackupURL string `json:"backupURL"`
+}
+
+// FoundationDBRestoreStatus describes the current status of the restore for a cluster.
+type FoundationDBRestoreStatus struct {
+	// Running describes whether the restore is currently running.
+	Running bool `json:"running,omitempty"`
+}
